add deploy file, move 2.x flows to prefect2_only
